Protection against NPE in code generation workflow and validatoin
Closes eMoflon/emoflon-tool#326
