+++ org.moflon.ide.ui/src/org/moflon/ide/ui/admin/handlers/AbstractValidateHandler.java
@@ -12,16 +12,23 @@ import org.eclipse.core.resources.IProject;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IAdaptable;
 import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.NullProgressMonitor;
+import org.eclipse.core.runtime.OperationCanceledException;
 import org.eclipse.core.runtime.SubMonitor;
+import org.eclipse.core.runtime.jobs.Job;
+import org.eclipse.core.runtime.jobs.JobGroup;
 import org.eclipse.jdt.core.IJavaProject;
 import org.eclipse.jface.viewers.ISelection;
 import org.eclipse.jface.viewers.TreeSelection;
 import org.eclipse.ui.IWorkingSet;
+import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.handlers.HandlerUtil;
+import org.gervarro.eclipse.task.ITask;
 import org.moflon.core.ui.AbstractCommandHandler;
 import org.moflon.core.utilities.MoflonConventions;
 import org.moflon.ide.core.CoreActivator;
+import org.moflon.ide.ui.preferences.EMoflonPreferenceInitializer;
 
 /**
  * Implementation base for validation with and without dumping the intermediate
@@ -106,6 +113,39 @@ public abstract class AbstractValidateHandler extends AbstractCommandHandler {
 		if (CoreActivator.isMoflonProject(project)) {
 			final IFile ecoreFile = MoflonConventions.getDefaultEcoreFile(project);
 			validateFile(ecoreFile, monitor);
+		}
+	}
+
+	@SuppressWarnings("deprecation")
+	protected void runValidation(final ITask validationTask) {
+		final long validationTimeoutMillis = EMoflonPreferenceInitializer.getValidationTimeoutMillis();
+		final String validationTimeoutMessage = "Validation took longer than " + validationTimeoutMillis / 1000
+				+ "s. This could(!) mean that some of your patterns have no valid search plan. You may increase the timeout value using the eMoflon property page";
+	
+		final Job job = new Job(validationTask.getTaskName()) {
+			@Override
+			protected IStatus run(final IProgressMonitor monitor) {
+				return validationTask.run(monitor);
+			}
+		};
+	
+		final boolean runInBackground = PlatformUI.getPreferenceStore().getBoolean("RUN_IN_BACKGROUND");
+		if (!runInBackground) {
+			PlatformUI.getWorkbench().getProgressService().showInDialog(null, job);
+		}
+	
+		final JobGroup jobGroup = new JobGroup("Validation job group", 1, 1);
+		job.setJobGroup(jobGroup);
+		job.schedule();
+		try {
+			boolean completed = jobGroup.join(validationTimeoutMillis, new NullProgressMonitor());
+			if (!completed) {
+				job.getThread().stop();
+				logger.error(validationTimeoutMessage);
+				throw new OperationCanceledException(validationTimeoutMessage);
+			}
+		} catch (InterruptedException e) {
+			throw new OperationCanceledException(validationTimeoutMessage);
 		}
 	}
 }
